thesis_project/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
elasticsearch==7.17.2
elasticsearch-dsl==7.4.0
idna==3.3
imageio==2.17.0
joblib==1.1.0
kafka-python==2.0.2
katna==0.9.2
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
scenedetect==0.5.6.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.4.8
tqdm==4.64.0
urllib3==1.26.9
python-json-logger==2.0.2
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
python-snappy==0.6.1
